+++ tests/unit/conftest.py
@@ -1,9 +1,7 @@
 from unittest.mock import Mock
 
 import pytest
-import stevedore
 
-import tavern
 from tavern.util.strict_util import StrictLevel
 
 
@@ -23,20 +21,3 @@ def fix_example_includes():
     }
 
     return includes.copy()
-
-
-@pytest.fixture(scope="session", autouse=True)
-def set_plugins():
-    def extension(name, point):
-        return stevedore.extension.Extension(name, point, point, point)
-
-    tavern.plugins.load_plugins.plugins = [
-        extension(
-            "requests",
-            tavern._plugins.rest.tavernhook.TavernRestPlugin,
-        ),
-        extension(
-            "paho-mqtt",
-            tavern._plugins.mqtt.tavernhook,
-        ),
-    ]
